OAK-3841: Change return type of Document.getModCount() to Long (ported to 1.2)
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/branches/1.2@1725932 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-core/src/test/java/org/apache/jackrabbit/oak/plugins/document/MultiDocumentStoreTest.java
@@ -101,7 +101,7 @@ public class MultiDocumentStoreTest extends AbstractMultiDocumentStoreTest {
         NodeDocument nd2 = super.ds2.find(Collection.NODES, id);
         assertNotNull(nd1);
         assertNotNull(nd2);
-        long firstVersion = nd1.getModCount().longValue();
+        long firstVersion = nd1.getModCount();
         assertEquals(firstVersion, nd2.getModCount().longValue());
 
         // letTimeElapse();
@@ -162,14 +162,14 @@ public class MultiDocumentStoreTest extends AbstractMultiDocumentStoreTest {
 
         // only run test if DS supports modcount
         if (orig.getModCount() != null) {
-            long origMc = orig.getModCount().longValue();
+            long origMc = orig.getModCount();
 
             UpdateOp up2 = new UpdateOp(id, false);
             up2.set("_id", id);
             up2.increment("_foo", 1L);
             super.ds2.update(Collection.NODES, Collections.singletonList(id), up2);
             NodeDocument ds2doc = super.ds2.find(Collection.NODES, id);
-            long ds2Mc = ds2doc.getModCount().longValue();
+            long ds2Mc = ds2doc.getModCount();
             assertTrue("_modCount needs to be > " + origMc + " but was " + ds2Mc, ds2Mc > origMc);
 
             UpdateOp up1 = new UpdateOp(id, false);
@@ -178,7 +178,7 @@ public class MultiDocumentStoreTest extends AbstractMultiDocumentStoreTest {
             super.ds1.update(Collection.NODES, Collections.singletonList(id), up1);
 
             NodeDocument ds1doc = super.ds1.find(Collection.NODES, id);
-            long ds1Mc = ds1doc.getModCount().longValue();
+            long ds1Mc = ds1doc.getModCount();
             assertTrue("_modCount needs to be > " + ds2Mc + " but was " + ds1Mc, ds1Mc > ds2Mc);
         }
     }
@@ -238,7 +238,7 @@ public class MultiDocumentStoreTest extends AbstractMultiDocumentStoreTest {
                 try {
                     DocumentStore ds = stores.get(random.nextInt(stores.size()));
                     NodeDocument d = ds.find(Collection.NODES, id);
-                    long modCount = d.getModCount().longValue();
+                    long modCount = d.getModCount();
                     NodeDocument seen = docs.get(modCount);
                     if (seen == null) {
                         docs.put(modCount, d);
